analysis/pyproject.toml
[tool.poetry]
name = "analysis"
version = "0.1.0"
description = ""
authors = ["fidhalkotta <fidhalkotta@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9"
matplotlib = "^3.8.2"
numpy = "^1.26.2"
pandas = "^2.1.3"
numpy-stl = "^3.1.1"
scipy = "^1.12.0"
open3d = "^0.18.0"
scikit-learn = "^1.4.0"
pyyaml = "^6.0.1"
addict = "^2.4.0"
tqdm = "^4.66.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
